Simplify (hopefully) and make ready for prod build
Now that Miso latest is on Hackage, I should try to use it. The idea
here is to attempt "as little nix shenanigans as I think I can get away
with".

I've yanked out anything related to Miso-style for now. The Miso
element API has changed and I've not yet had the chance to look at
updating it. Also, I'm not really sure what exactly to do to make
Miso-style practical. It's not really fast enough to actually use ... I
wonder if there isn't a way to generate as much CSS as possible at
compile time (like type-of-html).

I tried (really I did) to get `jsaddle` working, I reached my time-limit
on trying Nix builds and kept to my silly ghcjs-dev-server cause it was
easy to do so. There are so many missing things:
* You don't see warnings ... anywhere
* It doesn't know anything about Cabal
* There's no indication that anything is happening during (re)compilation.